Ryddet i Controllerne, gjort klar for å skulle i to applikasjoner. Skrevet om relaterte personer-søket så det ikke jobber på klagebehandlingid, men på fnr
