+++ src/main/kotlin/no/nav/klage/oppgave/api/controller/KlagebehandlingSearchController.kt
@@ -4,31 +4,38 @@ import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import no.nav.klage.oppgave.api.mapper.KlagebehandlingListMapper
 import no.nav.klage.oppgave.api.mapper.KlagebehandlingerSearchCriteriaMapper
-import no.nav.klage.oppgave.api.view.FnrSearchResponse
-import no.nav.klage.oppgave.api.view.NameSearchResponse
-import no.nav.klage.oppgave.api.view.SearchPersonByFnrInput
-import no.nav.klage.oppgave.api.view.SearchPersonByNameInput
+import no.nav.klage.oppgave.api.view.*
+import no.nav.klage.oppgave.clients.pdl.PdlFacade
+import no.nav.klage.oppgave.clients.pdl.Sivilstand
 import no.nav.klage.oppgave.config.SecurityConfiguration.Companion.ISSUER_AAD
+import no.nav.klage.oppgave.domain.KlagebehandlingerSearchCriteria
+import no.nav.klage.oppgave.domain.kodeverk.TemaTilgjengeligeForEktefelle
 import no.nav.klage.oppgave.repositories.InnloggetSaksbehandlerRepository
+import no.nav.klage.oppgave.service.ElasticsearchService
 import no.nav.klage.oppgave.service.PersonsoekService
 import no.nav.klage.oppgave.service.SaksbehandlerService
 import no.nav.klage.oppgave.util.getLogger
 import no.nav.security.token.support.core.api.ProtectedWithClaims
+import org.springframework.core.env.Environment
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import java.time.LocalDate
 
 @RestController
 @RequestMapping("/search")
 @Api(tags = ["kabal-api"])
 @ProtectedWithClaims(issuer = ISSUER_AAD)
 class KlagebehandlingSearchController(
-    private val klagebehandlingMapper: KlagebehandlingListMapper,
+    private val klagebehandlingListMapper: KlagebehandlingListMapper,
     private val klagebehandlingerSearchCriteriaMapper: KlagebehandlingerSearchCriteriaMapper,
     private val innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository,
     private val saksbehandlerService: SaksbehandlerService,
-    private val personsoekService: PersonsoekService
+    private val personsoekService: PersonsoekService,
+    private val environment: Environment,
+    private val pdlFacade: PdlFacade,
+    private val elasticsearchService: ElasticsearchService
 ) {
 
     companion object {
@@ -46,7 +53,7 @@ class KlagebehandlingSearchController(
         val personSoekHits = personsoekService.fnrSearch(searchCriteria)
         val saksbehandler = innloggetSaksbehandlerRepository.getInnloggetIdent()
         val valgtEnhet = saksbehandlerService.findValgtEnhet(saksbehandler)
-        return klagebehandlingMapper.mapPersonSoekHitsToFnrSearchResponse(
+        return klagebehandlingListMapper.mapPersonSoekHitsToFnrSearchResponse(
             personSoekHits = personSoekHits,
             saksbehandler = saksbehandler,
             tilgangTilTemaer = valgtEnhet.temaer
@@ -67,6 +74,44 @@ class KlagebehandlingSearchController(
                     name = it.name
                 )
             }
+        )
+    }
+
+    @PostMapping("/relaterte")
+    fun getRelaterteKlagebehandlinger(
+        @RequestBody input: SearchPersonByFnrInput
+    ): KlagebehandlingerListRespons {
+        //TODO: Move logic to PersonsoekService
+        val lovligeTemaer =
+            saksbehandlerService.findValgtEnhet(innloggetSaksbehandlerRepository.getInnloggetIdent()).temaer
+        val sivilstand: Sivilstand? = pdlFacade.getPersonInfo(input.query).sivilstand
+
+        val searchCriteria = KlagebehandlingerSearchCriteria(
+            statuskategori = KlagebehandlingerSearchCriteria.Statuskategori.ALLE,
+            ferdigstiltFom = LocalDate.now().minusMonths(12),
+            foedselsnr = listOf(input.query),
+            extraPersonAndTema = sivilstand?.let {
+                KlagebehandlingerSearchCriteria.ExtraPersonAndTema(
+                    foedselsnr = it.foedselsnr,
+                    temaer = TemaTilgjengeligeForEktefelle.temaerTilgjengeligForEktefelle(environment).toList()
+                )
+            },
+            offset = 0,
+            limit = 100,
+            projection = KlagebehandlingerSearchCriteria.Projection.UTVIDET,
+        )
+
+        val esResponse = elasticsearchService.findByCriteria(searchCriteria)
+        return KlagebehandlingerListRespons(
+            antallTreffTotalt = esResponse.totalHits.toInt(),
+            klagebehandlinger = klagebehandlingListMapper.mapEsKlagebehandlingerToListView(
+                esKlagebehandlinger = esResponse.searchHits.map { it.content },
+                viseUtvidet = true,
+                viseFullfoerte = true,
+                saksbehandler = searchCriteria.saksbehandler,
+                tilgangTilTemaer = lovligeTemaer,
+                sivilstand = sivilstand
+            )
         )
     }
 }
